heatmap provider

+++ island/src/main/java/ca/mcmaster/cas/se2aa4/a3/island/shape/Shape.java
@@ -1,5 +1,5 @@
 package ca.mcmaster.cas.se2aa4.a3.island.shape;
 
 public interface Shape {
-    boolean contains(int x, int y);
+    boolean contains(double x, double y);
 }
